backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.4
scikit-learn==1.4.1.post1
xgboost==2.0.3
transformers==4.38.2
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2+cpu
shap==0.45.0
httpx==0.27.0
python-whois==0.8.0
beautifulsoup4==4.12.3
motor==3.4.0
crewai==0.28.8
langchain-groq==0.1.3
langchain-openai==0.1.6
python-dotenv==1.0.1
datasets==2.18.0
accelerate==0.28.0
